[Feat] Change Emacs config to work cohesively on main repository
This commit changes the repository to work cohesively with the master repository
from sanity inc. Instead of replacing files or storing their modifications in
this repository, I am modifying only what's needed and adding only what's needed
here. This allows me to not need to fork his repo and also have the benefits of
future updates.